refactor : ClinicClientPage 페이지 구조 변경 (#96)
ScheduleSection에서 ScheduleList로 변경
