+++ src/features/clinic/ui/ClinicClientPage.tsx
@@ -1,11 +1,13 @@
 'use client'
 
-import { BottomNav, Calendar, Profile } from '@/components'
+import { useRouter } from 'next/navigation'
+
+import { BottomNav, Calendar, Label, Profile } from '@/components'
+import ActionTag from '@/components/actionTag/ActionTag'
+import { ScheduleList } from '@/features'
 import { useUserStore } from '@/store'
 
 import { useScheduleList } from '../service/useScheduleList'
-
-import { ScheduleSection } from './ScheduleSection'
 
 type ClinicClientPageProps = {
   userId: string
@@ -13,7 +15,12 @@ type ClinicClientPageProps = {
 
 export const ClinicClientPage = ({ userId }: ClinicClientPageProps) => {
   const { user } = useUserStore()
+  const router = useRouter()
   const scheduleList = useScheduleList(userId)
+
+  const handleClickPlusButton = () => {
+    router.push(`/clinic/${userId}/write`)
+  }
 
   return (
     <>
@@ -26,7 +33,13 @@ export const ClinicClientPage = ({ userId }: ClinicClientPageProps) => {
           <Calendar />
         </section>
 
-        <ScheduleSection userId={userId} scheduleList={scheduleList} />
+        <section className="mt-7">
+          <div className="flex-between mb-3">
+            <Label icon="clinic-label">병원 내원 일정</Label>
+            <ActionTag.Plus label="추가" primary onClick={handleClickPlusButton} />
+          </div>
+          <ScheduleList userId={userId} scheduleList={scheduleList} />
+        </section>
       </main>
       <BottomNav />
     </>
