+++ src/features/index.ts
@@ -17,5 +17,5 @@ export * from './clinic/ui/ClinicInfoModal'
 export * from './clinic/ui/ClinicWriteClientPage'
 export * from './clinic/ui/DateBottomSheet'
 export * from './clinic/ui/ScheduleItem'
-export * from './clinic/ui/ScheduleSection'
+export * from './clinic/ui/ScheduleList'
 export * from './clinic/ui/TagBottomSheet'
